+++ ouroboros-network-api/src/Cardano/Network/ConsensusMode.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric  #-}
 
-module Cardano.Node.ConsensusMode where
+module Cardano.Network.ConsensusMode where
 
 import Data.Aeson
 import GHC.Generics (Generic)
