+++ ouroboros-network/src/Cardano/Network/PeerSelection/PeerChurnArgs.hs
@@ -1,9 +1,9 @@
-module Cardano.Node.PeerSelection.PeerChurnArgs where
+module Cardano.Network.PeerSelection.PeerChurnArgs where
 
-import Cardano.Node.ArgumentsExtra (ConsensusModePeerTargets)
-import Cardano.Node.ConsensusMode (ConsensusMode)
-import Cardano.Node.PeerSelection.Bootstrap (UseBootstrapPeers)
-import Cardano.Node.PeerSelection.Types (ChurnMode)
+import Cardano.Network.ArgumentsExtra (ConsensusModePeerTargets)
+import Cardano.Network.ConsensusMode (ConsensusMode)
+import Cardano.Network.PeerSelection.Bootstrap (UseBootstrapPeers)
+import Cardano.Network.PeerSelection.Types (ChurnMode)
 import Control.Concurrent.Class.MonadSTM.Strict
 import Ouroboros.Network.BlockFetch.ConsensusInterface (FetchMode)
