+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/Governor/KnownPeers.hs
@@ -26,10 +26,10 @@ import Control.Monad.Class.MonadSTM
 import Control.Monad.Class.MonadTime.SI
 import Control.Monad.Class.MonadTimer.SI
 
-import Cardano.Node.PeerSelection.Bootstrap (requiresBootstrapPeers)
-import Cardano.Node.PeerSelection.Governor.PeerSelectionState
+import Cardano.Network.PeerSelection.Bootstrap (requiresBootstrapPeers)
+import Cardano.Network.PeerSelection.Governor.PeerSelectionState
            (CardanoPeerSelectionState (..))
-import Cardano.Node.PublicRootPeers (CardanoPublicRootPeers)
+import Cardano.Network.PublicRootPeers (CardanoPublicRootPeers)
 import Ouroboros.Network.Diffusion.Policies qualified as Policies
 import Ouroboros.Network.PeerSelection.Governor.Types
 import Ouroboros.Network.PeerSelection.PeerAdvertise (PeerAdvertise (..))
