+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/PublicRootPeers.hs
@@ -34,8 +34,8 @@ module Ouroboros.Network.PeerSelection.PublicRootPeers
 
 import Prelude hiding (null)
 
-import Cardano.Node.PublicRootPeers (CardanoPublicRootPeers (..))
-import Cardano.Node.PublicRootPeers qualified as CPRP
+import Cardano.Network.PublicRootPeers (CardanoPublicRootPeers (..))
+import Cardano.Network.PublicRootPeers qualified as CPRP
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
 import Data.Set (Set, (\\))
